requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Redis and Celery
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# OpenAI and AI libraries
openai==1.3.7
mistralai==0.0.12

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.2

# Additional utilities
schedule==1.2.0
aiofiles==23.2.1
jinja2==3.1.2
